% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{getMsgTracer}
\alias{getMsgTracer}
\title{A Message Tracer}
\usage{
getMsgTracer()
}
\value{
a new tracer instance
}
\description{
This implements a tracer that delegates function calls to \code{message()}.
}
\examples{
z <- ot::getMsgTracer()
}
