% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{getMockTracer}
\alias{getMockTracer}
\alias{clearMockTracer}
\title{A Mock Tracer implementation}
\usage{
getMockTracer()

clearMockTracer(tracer)
}
\arguments{
\item{tracer}{the mock tracer}
}
\description{
This implements a Tracer and Span which stores all function arguments in an environment.
}
