% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noop.R
\name{getNoOpTracer}
\alias{getNoOpTracer}
\title{A NoOp Tracer implementation}
\usage{
getNoOpTracer()
}
\value{
a tracer instance
}
\description{
The OpenTracing specification mandates a tracer implementation which does nothing.
}
\examples{
z <- ot::getNoOpTracer()
}
