% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ot.R
\docType{package}
\name{ot}
\alias{ot}
\title{Open Tracing}
\description{
Open Tracing is a specification for trace logging. This package provides generics for the required methods, and a minimal
implementation for testing purposes.
}
\seealso{
\url{https://opentracing.io/}
}
