% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{getMockTracer}
\alias{getMockTracer}
\title{A Mock Tracer implementation}
\usage{
getMockTracer()
}
\value{
a new tracer instance
}
\description{
This implements a Tracer and Span which stores all function arguments in an environment.
}
\examples{
z <- ot::getMockTracer()
}
